Drop support for JSON-based config files
