+++ metaphor/common/cli.py
@@ -1,6 +1,5 @@
 import argparse
 import logging
-import os
 from typing import Type
 
 from .extractor import BaseExtractor
@@ -17,12 +16,4 @@ def cli_main(description: str, extractor_cls: Type[BaseExtractor]):
 
     extractor = extractor_cls()
     config_cls = extractor.config_class()
-
-    extension = os.path.splitext(args.config)[-1].lower()
-
-    if extension == ".json":
-        extractor.run(config=config_cls.from_json_file(args.config))
-    elif extension == ".yml" or extension == ".yaml":
-        extractor.run(config=config_cls.from_yaml_file(args.config))
-    else:
-        raise ValueError(f"Unknown config file extension: {extension}")
+    extractor.run(config=config_cls.from_yaml_file(args.config))
