+++ metaphor/dbt/config.py
@@ -3,12 +3,12 @@ from typing import Optional
 
 from serde import deserialize
 
-from metaphor.common.extractor import RunConfig
+from metaphor.common.base_config import BaseConfig
 
 
 @deserialize
 @dataclass
-class DbtRunConfig(RunConfig):
+class DbtRunConfig(BaseConfig):
     manifest: str
     catalog: Optional[str] = None
