+++ metaphor/metabase/config.py
@@ -2,12 +2,12 @@ from dataclasses import dataclass
 
 from serde import deserialize
 
-from metaphor.common.extractor import RunConfig
+from metaphor.common.base_config import BaseConfig
 
 
 @deserialize
 @dataclass
-class MetabaseRunConfig(RunConfig):
+class MetabaseRunConfig(BaseConfig):
     server_url: str
     username: str
     password: str
