+++ metaphor/postgresql/config.py
@@ -3,13 +3,13 @@ from typing import Optional
 
 from serde import deserialize
 
-from metaphor.common.extractor import RunConfig
+from metaphor.common.base_config import BaseConfig
 from metaphor.common.filter import DatasetFilter
 
 
 @deserialize
 @dataclass
-class PostgreSQLRunConfig(RunConfig):
+class PostgreSQLRunConfig(BaseConfig):
     host: str
     database: str
     user: str
