--- tests/bigquery/config.json
@@ -1,6 +1,0 @@
-{
-  "key_path": "key_path",
-  "project_id": "project_id",
-  "dataset_filters": ["f1", "f2"],
-  "output": null
-}
