--- tests/bigquery/profile/config.json
@@ -1,5 +1,0 @@
-{
-  "key_path": "key_path",
-  "project_id": "project_id",
-  "output": null
-}
