+++ tests/bigquery/profile/test_config.py
@@ -3,18 +3,6 @@ from metaphor.common.filter import DatasetFilter
 from metaphor.common.sampling import SamplingConfig
 
 
-def test_json_config(test_root_dir):
-    config = BigQueryProfileRunConfig.from_json_file(
-        f"{test_root_dir}/bigquery/profile/config.json"
-    )
-
-    assert config == BigQueryProfileRunConfig(
-        key_path="key_path",
-        project_id="project_id",
-        output=None,
-    )
-
-
 def test_yaml_config(test_root_dir):
     config = BigQueryProfileRunConfig.from_yaml_file(
         f"{test_root_dir}/bigquery/profile/config.yml"
