+++ tests/bigquery/test_config.py
@@ -1,17 +1,6 @@
 from metaphor.bigquery.extractor import BigQueryRunConfig
 
 
-def test_json_config(test_root_dir):
-    config = BigQueryRunConfig.from_json_file(f"{test_root_dir}/bigquery/config.json")
-
-    assert config == BigQueryRunConfig(
-        key_path="key_path",
-        project_id="project_id",
-        dataset_filters=["f1", "f2"],
-        output=None,
-    )
-
-
 def test_yaml_config(test_root_dir):
     config = BigQueryRunConfig.from_yaml_file(f"{test_root_dir}/bigquery/config.yml")
