+++ tests/bigquery/usage/test_config.py
@@ -1,21 +1,6 @@
 from metaphor.bigquery.usage.config import BigQueryUsageRunConfig
 
 
-def test_json_config(test_root_dir):
-    config = BigQueryUsageRunConfig.from_json_file(
-        f"{test_root_dir}/bigquery/usage/config.json"
-    )
-
-    assert config == BigQueryUsageRunConfig(
-        key_path="key_path",
-        project_id="project_id",
-        dataset_filters=["f1", "f2"],
-        lookback_days=365,
-        batch_size=100,
-        output=None,
-    )
-
-
 def test_yaml_config(test_root_dir):
     config = BigQueryUsageRunConfig.from_yaml_file(
         f"{test_root_dir}/bigquery/usage/config.yml"
