--- tests/common/config.json
@@ -1,14 +1,0 @@
-{
-  "output": {
-    "api": {
-      "url": "url",
-      "api_key": "api_key",
-      "batch_size": 1,
-      "timeout": 2
-    },
-    "file": {
-      "assume_role_arn": "arn",
-      "directory": "path"
-    }
-  }
-}
