--- tests/dbt/config.json
@@ -1,5 +1,0 @@
-{
-  "manifest": "manifest",
-  "catalog": "catalog",
-  "output": null
-}
