--- tests/google_directory/config.json
@@ -1,4 +1,0 @@
-{
-  "token_file": "token_file",
-  "output": null
-}
