--- tests/looker/config.json
@@ -1,16 +1,0 @@
-{
-  "base_url": "base_url",
-  "client_id": "client_id",
-  "client_secret": "client_secret",
-  "lookml_dir": "lookml_dir",
-  "connections": {
-    "conn1": {
-      "database": "db1",
-      "default_schema": "schema1",
-      "account": "account1"
-    }
-  },
-  "verify_ssl": true,
-  "timeout": 1,
-  "output": null
-}
