--- tests/metabase/config.json
@@ -1,6 +1,0 @@
-{
-  "server_url": "https://metaphor.metabaseapp.com",
-  "username": "yi@metaphor.io",
-  "password": "xyz",
-  "output": null
-}
