--- tests/postgresql/config.json
@@ -1,7 +1,0 @@
-{
-  "host": "host",
-  "database": "database",
-  "user": "user",
-  "password": "password",
-  "output": null
-}
