--- tests/snowflake/config.json
@@ -1,7 +1,0 @@
-{
-  "account": "account",
-  "user": "user",
-  "password": "password",
-  "default_database": "database",
-  "output": null
-}
