--- tests/tableau/config.json
@@ -1,9 +1,0 @@
-{
-  "server_url": "https://10ax.online.tableau.com/",
-  "site_name": "abc",
-  "user_password": {
-    "username": "yi@metaphor.io",
-    "password": "xyz"
-  },
-  "output": null
-}
